backend/requirements.txt
# Dependencias del Backend - ColorMetría
# Framework web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Procesamiento de imágenes
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.26.0

# Detección facial y landmarks
mediapipe>=0.10.8

# Análisis de colores
scikit-learn>=1.3.0
scipy>=1.11.0
pandas>=2.0.0
joblib>=1.3.0

# Utilidades
python-dotenv>=1.0.0
pydantic>=2.5.0

# CORS
starlette>=0.32.0

# Base de datos
SQLAlchemy>=2.0.0
aiosqlite>=0.19.0

# Opcional: atributos faciales (edad, género, emoción) vía DeepFace.
# Aumenta peso y tiempo de arranque (TensorFlow). Sin instalar, la API funciona igual.
# deepface>=0.0.92
